[AEA-4929] review misc cleanup
