+++ .github/workflows/cdk_package_code.yml
@@ -1,4 +1,3 @@
-# TODO: make the pipelines use this; ci, pull_request, release
 name: cdk package code
 
 on:
