+++ .github/workflows/create_release_notes.yml
@@ -1,4 +1,3 @@
-# TODO: use this
 name: create release notes
 
 on:
